backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
<<<<<<< HEAD
openai==0.28
=======
openai==1.3.7
>>>>>>> 7064d9fe2673553a70d9f990ad4b7a1ae5e6e69b
google-api-python-client==2.108.0
certifi==2023.11.17
lxml==4.9.3
Pillow==10.1.0
reportlab==4.0.7
brotli==1.1.0
python-multipart==0.0.6
jinja2==3.1.2
# Lock sub-dependencies
charset-normalizer==3.3.2
idna==3.6
multidict==6.0.4
yarl==1.9.4
<<<<<<< HEAD
soupsieve==2.5 
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
pydantic==2.5.0
python-multipart==0.0.6
redis==5.0.1
requests==2.31.0
lxml==4.9.3
python-dotenv==1.0.0
=======
soupsieve==2.5
>>>>>>> 7064d9fe2673553a70d9f990ad4b7a1ae5e6e69b
